test: Replace unsigned integer types with signed equivalents
Replace all uses of uint8, uint16, uint32, uint64 with their signed
counterparts in test schemas, following the removal of unsigned types
from core types.

Changes:
- uint8 → int16 (larger signed type for unsigned 8-bit range)
- uint16 → int32 (larger signed type for unsigned 16-bit range)
- uint32 → int64 (larger signed type for unsigned 32-bit range)
- int unsigned → int64 (native type replacement)

Updated files:
- tests/schema_simple.py: E.M.id_m
- tests/schema_type_aliases.py: Remove unsigned types from test table
- tests/schema.py: Auto.id, Ephys.Channel.channel
- tests/schema_university.py: Student.student_id, Course.course
- tests/integration/test_type_aliases.py: Remove unsigned type tests
- tests/integration/test_hidden_job_metadata.py: All uint8 → int16

Note: test_blob.py and test_blob_matlab.py unchanged (testing numpy
dtypes in serialization, not DataJoint table definitions).

Part of v2.0 core type system revision.

+++ tests/schema.py
@@ -67,7 +67,7 @@ class TTestNoExtra(dj.Manual):
 
 class Auto(dj.Lookup):
     definition = """
-    id  : uint8
+    id  : int16
     ---
     name :varchar(12)
     """
@@ -195,7 +195,7 @@ class Ephys(dj.Imported):
     class Channel(dj.Part):
         definition = """     # subtable containing individual channels
         -> master
-        channel    :uint8   # channel number within Ephys
+        channel    :int16   # channel number within Ephys
         ----
         voltage    : <blob>
         current = null : <blob>   # optional current to test null handling
